Now all the problems are fixed
we fixed the problem of two skeleton robosystem
we fixed the problem where robotbuilder was going to differents proyects
we put LEDS
we descomment a part of code that make the drivetrain go faster, by making the victors follow the talons.
we fixed a two problems with quick turn now it's working , the problem was a conflict between commands that was  making a competing drive command.